docker/requirements.txt
gluoncv
flask
gunicorn
boto3
gevent
Pillow